NN-2079 initial controller and styling
