+++ static/stylesheets/application.css
@@ -5903,6 +5903,14 @@
   color: white;
 }
 
+.padding-right {
+  padding-right: 17px;
+}
+
+.margin-bottom {
+  margin-bottom: 30px;
+}
+
 /*HEADER STYLES */
 .right-content {
   float: right;
@@ -6011,6 +6019,10 @@
   max-width: 1170px !important;
 }
 
+.govuk-grid-column-one-quarter {
+  padding: 0 15px 15px !important;
+}
+
 #menuToggle,
 #menuItems {
   display: none;
@@ -6022,4 +6034,85 @@
 
 #menuToggle:checked ~ .right-menu-info-wrap:before {
   transform: rotate(180deg);
+}
+
+.pure-u-md-1,
+.pure-u-md-1-1,
+.pure-u-md-1-2,
+.pure-u-md-1-3,
+.pure-u-md-2-3,
+.pure-u-md-1-4,
+.pure-u-md-3-4,
+.pure-u-md-1-6,
+.pure-u-md-5-6,
+.pure-u-md-1-12,
+.pure-u-md-2-12,
+.pure-u-md-3-12,
+.pure-u-md-4-12,
+.pure-u-md-5-12,
+.pure-u-md-6-12,
+.pure-u-md-7-12,
+.pure-u-md-8-12,
+.pure-u-md-9-12,
+.pure-u-md-10-12,
+.pure-u-md-11-12,
+.pure-u-md-12-12 {
+  display: inline-block;
+  *display: inline;
+  zoom: 1;
+  letter-spacing: normal;
+  word-spacing: normal;
+  vertical-align: top;
+  text-rendering: auto;
+}
+
+@media screen and (min-width: 48em) {
+  .pure-u-md-1-6,
+  .pure-u-md-4-24 {
+    width: 16.6667%;
+  }
+}
+
+select.form-control {
+  width: 100%;
+  height: 2em;
+}
+
+.form-control {
+  -webkit-box-sizing: border-box;
+  -moz-box-sizing: border-box;
+  box-sizing: border-box;
+  font-family: 'nta', Arial, sans-serif;
+  font-weight: 400;
+  text-transform: none;
+  font-size: 16px;
+  line-height: 1.25;
+  width: 100%;
+  padding: 5px 4px 4px;
+  border: 2px solid #0b0c0c;
+}
+
+@media (min-width: 641px) {
+  .form-control {
+    font-size: 19px;
+    line-height: 1.31579;
+  }
+}
+
+.form-label {
+  font-family: 'nta', Arial, sans-serif;
+  font-weight: 400;
+  text-transform: none;
+  font-size: 16px;
+  line-height: 1.25;
+  display: block;
+  color: #0b0c0c;
+  padding-bottom: 2px;
+}
+
+@media (min-width: 641px) {
+  .form-label {
+    font-size: 19px;
+    line-height: 1.31579;
+  }
 }
