+++ views/components/datePicker.njk
@@ -1,0 +1,8 @@
+{% from "govuk/components/input/macro.njk" import govukInput %}
+
+<div class="pure-u-md-1-6 padding-right margin-bottom">
+<label class="form-label" for="date-picker">
+    Date
+</label>
+  <input class="govuk-input" id="event-name" name="date-picker" type="text">
+</div>
