+++ views/components/periodSelect.njk
@@ -1,0 +1,24 @@
+
+<div class="pure-u-md-1-6">
+<label class="form-label" for="period-select">
+    Period
+</label>
+
+<select
+    id="period-select"
+    name="period-select"
+    class="form-control"
+    value={period}
+    onChange={handlePeriodChange}
+>
+    <option key="MORNING" value="AM">
+    Morning (AM)
+    </option>
+    <option key="AFTERNOON" value="PM">
+    Afternoon (PM)
+    </option>
+    <option key="EVENING" value="ED">
+    Evening (ED)
+    </option>
+</select>
+</div>
